+++ filters/miditranspose.c
@@ -14,8 +14,8 @@ MFD_FILTER(miditranspose)
 #elif defined MX_CODE
 
 static void filter_preproc_miditranspose(MidiFilter* self) {
-	if (floor(self->lcfg[1]) == floor(*self->cfg[1])) return;
-	const int transp = rint(*(self->cfg[1]));
+	if (floorf(self->lcfg[1]) == floorf(*self->cfg[1])) return;
+	const int transp = rintf(*(self->cfg[1]));
 
 	int c,k;
 	uint8_t buf[3];
@@ -65,8 +65,8 @@ filter_midi_miditranspose(MidiFilter* self,
 		const uint8_t* const buffer,
 		uint32_t size)
 {
-	const int chs = midi_limit_chn(floor(*self->cfg[0]) -1);
-	const int transp = rint(*(self->cfg[1]));
+	const int chs = midi_limit_chn(floorf(*self->cfg[0]) -1);
+	const int transp = rintf(*(self->cfg[1]));
 
 	const uint8_t chn = buffer[0] & 0x0f;
 	const uint8_t key = buffer[1] & 0x7f;
@@ -76,7 +76,7 @@ filter_midi_miditranspose(MidiFilter* self,
 
 	if (size != 3
 			|| !(mst == MIDI_NOTEON || mst == MIDI_NOTEOFF || mst == MIDI_POLYKEYPRESSURE)
-			|| !(floor(*self->cfg[0]) == 0 || chs == chn)
+			|| !(floorf(*self->cfg[0]) == 0 || chs == chn)
 		 )
 	{
 		forge_midimessage(self, tme, buffer, size);
