+++ filters/notetocc.c
@@ -34,13 +34,13 @@ filter_midi_notetocc(MidiFilter* self,
 		const uint8_t* const buffer,
 		uint32_t size)
 {
-	const uint8_t chs = midi_limit_chn(floor(*self->cfg[0]) -1);
+	const uint8_t chs = midi_limit_chn(floorf(*self->cfg[0]) -1);
 	const uint8_t chn = buffer[0] & 0x0f;
 	const uint8_t mst = buffer[0] & 0xf0;
 
 	if (size != 3
 			|| !(mst == MIDI_NOTEON || mst == MIDI_NOTEOFF)
-			|| !(floor(*self->cfg[0]) == 0 || chs == chn)
+			|| !(floorf(*self->cfg[0]) == 0 || chs == chn)
 		 )
 	{
 		forge_midimessage(self, tme, buffer, size);
@@ -49,7 +49,7 @@ filter_midi_notetocc(MidiFilter* self,
 
 	const uint8_t key = buffer[1] & 0x7f;
 	const uint8_t vel = buffer[2] & 0x7f;
-	const int mode = RAIL(floor(*self->cfg[1]),0, 3);
+	const int mode = RAIL(floorf(*self->cfg[1]),0, 3);
 
 	const uint8_t param = midi_limit_val(floorf(*self->cfg[1]));
 	const uint8_t kfltr = midi_limit_val(floorf(*self->cfg[2]));
